# target/classes/META-INF/swagger/actor-0.0.yml
openapi: 3.0.1
info:
  title: Actor
  version: "0.0"
paths:
  /actors:
    get:
      summary: Actors list
      description: desc actor list
      operationId: actor_list
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
    post:
      operationId: addActor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Actor'
        required: true
      responses:
        "200":
          description: addActor 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
  /actors/selected/{ids}:
    get:
      operationId: getActorsByIds
      parameters:
      - name: ids
        in: path
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int64
      responses:
        "200":
          description: getActorsByIds 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Actor'
  /actors/{id}:
    get:
      operationId: getActor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: getActor 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
    delete:
      operationId: deleteActor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: deleteActor 200 response
          content:
            application/json:
              schema:
                type: object
components:
  schemas:
    Actor:
      required:
      - firstName
      - lastName
      - oscarPrized
      - rating
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        rating:
          type: integer
          format: int32
        oscarPrized:
          type: boolean